# api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "梭子 - 智能日程管理平台 API",
    "version": "1.0.0",
    "description": "梭子（Shuttle）是一个智能日程记录与管理平台，提供高效记录、AI辅助规划、积分激励与个性化装扮等功能。\n\n## 微服务架构\n- **user-service**: 用户注册登录、信息管理、积分账户\n- **schedule-service**: 日程CRUD、计时控制、日历视图、统计接口\n- **ai-service**: AI规划生成、每日/月度总结\n- **notification-service**: 消息推送、提醒生成\n- **mall-service**: 虚拟商品管理、购买记录、用户装扮\n\n## 认证方式\n除登录注册外，所有接口需在 Header 中携带 JWT Token：\n```\nAuthorization: Bearer <token>\n```",
    "contact": {
      "name": "梭子开发团队"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Gateway 统一入口 (开发环境)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token 认证，登录接口返回"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码，200 表示成功",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "提示信息",
            "example": "操作成功"
          },
          "data": {
            "description": "响应数据"
          }
        }
      },
      "PageResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "查询成功"
          },
          "data": {
            "type": "object",
            "properties": {
              "records": {
                "type": "array",
                "description": "数据列表"
              },
              "total": {
                "type": "integer",
                "description": "总记录数"
              },
              "page": {
                "type": "integer",
                "description": "当前页码"
              },
              "pageSize": {
                "type": "integer",
                "description": "每页大小"
              },
              "totalPages": {
                "type": "integer",
                "description": "总页数"
              }
            }
          }
        }
      },
...
      "Schedule": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "description": "日程标题",
            "example": "学习Spring Boot"
          },
          "description": {
            "type": "string",
            "description": "详细描述",
            "example": "完成第3章：REST API开发"
          },
          "plannedStartTime": {
            "type": "string",
            "format": "date-time",
            "description": "计划开始时间"
          },
          "plannedDuration": {
            "type": "integer",
            "description": "预计时长（分钟）",
            "example": 120
          },
          "actualStartTime": {
            "type": "string",
            "format": "date-time",
            "description": "实际开始时间"
          },
          "actualEndTime": {
            "type": "string",
            "format": "date-time",
            "description": "实际结束时间"
          },
          "actualDuration": {
            "type": "integer",
            "description": "实际持续时长（秒）"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "IN_PROGRESS",
              "PAUSED",
              "COMPLETED",
              "CANCELLED",
              "TIMEOUT"
            ],
            "description": "日程状态：PENDING=待开始, IN_PROGRESS=进行中, PAUSED=已暂停, COMPLETED=已完成, CANCELLED=已取消, TIMEOUT=已超时"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ],
            "description": "优先级",
            "default": "MEDIUM"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "标签列表",
            "example": [
              "学习",
              "Java"
            ]
          },
          "color": {
            "type": "string",
            "description": "日程显示颜色（十六进制）",
            "example": "#4A90D9"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateScheduleRequest": {
        "type": "object",
        "required": [
          "title",
          "plannedStartTime"
        ],
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "description": "日程标题"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "description": "详细描述"
          },
          "plannedStartTime": {
            "type": "string",
            "format": "date-time",
            "description": "计划开始时间"
          },
          "plannedDuration": {
            "type": "integer",
            "minimum": 1,
            "description": "预计时长（分钟），默认60",
            "default": 60
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ],
            "description": "优先级",
            "default": "MEDIUM"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "标签列表"
          },
          "color": {
            "type": "string",
            "description": "显示颜色"
          }
        }
      },
      "UpdateScheduleRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 1000
          },
          "plannedStartTime": {
            "type": "string",
            "format": "date-time"
          },
          "plannedDuration": {
            "type": "integer",
            "minimum": 1
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "color": {
            "type": "string"
          }
        }
      },
      "TimeRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "scheduleId": {
            "type": "integer",
            "format": "int64"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "duration": {
            "type": "integer",
            "description": "该段时长（秒）"
          },
          "type": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "PAUSE"
            ],
            "description": "计时类型"
          }
        }
      },
...
      "BatchDeleteRequest": {
        "type": "object",
        "required": [
          "ids"
        ],
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "description": "要删除的 id 列表（去重后 ≤ 200）"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/schedules": {
      "post": {
        "tags": [
          "日程接口"
        ],
        "summary": "创建日程",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "createSchedule",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Schedule"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "日程接口"
        ],
        "summary": "查询日程列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "listSchedules",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "指定日期查询"
          },
          {
            "name": "view",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "day",
                "week",
                "month"
              ]
            },
            "description": "视图类型"
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "开始日期"
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "结束日期"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "状态筛选，多个用逗号分隔"
          },
          {
            "name": "priority",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "优先级筛选"
          },
          {
            "name": "tag",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "标签筛选"
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "标题/描述关键词搜索"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "plannedStartTime",
                "priority",
                "createdAt"
              ]
            },
            "description": "排序字段"
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResult"
                }
              }
            }
          }
        }
      }
    },
    "/schedules/{id}": {
      "get": {
        "tags": [
          "日程接口"
        ],
        "summary": "查询日程详情",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "getSchedule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/Schedule"
                            },
                            {
                              "properties": {
                                "timeRecords": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/TimeRecord"
                                  }
                                }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "日程不存在"
          }
        }
      },
      "put": {
        "tags": [
          "日程接口"
        ],
        "summary": "更新日程",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "updateSchedule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Schedule"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "日程接口"
        ],
        "summary": "删除日程",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "deleteSchedule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          },
          "404": {
            "description": "日程不存在"
          }
        }
      }
    },
...
    "/schedules/active": {
      "get": {
        "tags": [
          "日程接口"
        ],
        "summary": "获取当前进行中的日程",
        "description": "返回用户当前正在计时或暂停的日程（用于计时器持久化恢复）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "getActiveSchedule",
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "nullable": true,
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/Schedule"
                            },
                            {
                              "properties": {
                                "elapsedSeconds": {
                                  "type": "integer",
                                  "description": "当前已计时秒数"
                                },
                                "timerStartedAt": {
                                  "type": "string",
                                  "format": "date-time",
                                  "description": "最近一次计时开始时间"
                                }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/schedules/batch": {
      "post": {
        "tags": [
          "日程接口"
        ],
        "summary": "批量创建日程",
        "description": "配合 AI 规划，一次性创建多个日程",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "batchCreateSchedules",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "schedules"
                ],
                "properties": {
                  "schedules": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/CreateScheduleRequest"
                    },
                    "maxItems": 50
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "批量创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Schedule"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/schedules/batch-delete": {
      "post": {
        "tags": [
          "日程接口"
        ],
        "summary": "批量删除日程",
        "description": "批量软删除，事务保证全部成功或全部失败；包含不存在或无权限的 id 时整体回滚（404）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "batchDeleteSchedules",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "integer",
                          "description": "删除条数"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "ids 为空或超过 200 个"
          },
          "404": {
            "description": "部分日程不存在或无权删除，已整体回滚"
          }
        }
      }
    },
...
  }
}